requirements.txt
openai
google-genai
tqdm
